refactor: separate logic into new helper class
References: https://outsystemsrd.atlassian.net/browse/RMET-3905
