+++ packages/android-lib/src/main/kotlin/com/outsystems/plugins/osgeolocation/controller/OSGeolocationController.kt
@@ -5,22 +5,13 @@ import android.content.Context
 import android.location.Location
 import android.location.LocationManager
 import android.os.Build
-import android.os.Looper
 import android.util.Log
 import androidx.activity.result.ActivityResultLauncher
 import androidx.activity.result.IntentSenderRequest
 import androidx.core.location.LocationManagerCompat
-import com.google.android.gms.common.ConnectionResult
-import com.google.android.gms.common.GoogleApiAvailability
-import com.google.android.gms.common.api.ResolvableApiException
-import com.google.android.gms.location.CurrentLocationRequest
 import com.google.android.gms.location.FusedLocationProviderClient
 import com.google.android.gms.location.LocationCallback
-import com.google.android.gms.location.LocationRequest
 import com.google.android.gms.location.LocationResult
-import com.google.android.gms.location.LocationServices
-import com.google.android.gms.location.LocationSettingsRequest
-import com.google.android.gms.location.Priority
 import com.outsystems.plugins.osgeolocation.model.OSLocationException
 import com.outsystems.plugins.osgeolocation.model.OSLocationOptions
 import com.outsystems.plugins.osgeolocation.model.OSLocationResult
@@ -29,15 +20,18 @@ import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.callbackFlow
 import kotlinx.coroutines.flow.first
-import kotlinx.coroutines.tasks.await
 
 /**
  * Entry point in OSGeolocationLib-Android
  *
  */
 class OSGeolocationController(
-    private val fusedLocationClient: FusedLocationProviderClient,
-    private val activityLauncher: ActivityResultLauncher<IntentSenderRequest>
+    fusedLocationClient: FusedLocationProviderClient,
+    activityLauncher: ActivityResultLauncher<IntentSenderRequest>,
+    private val helper: OSGeolocationServiceHelper = OSGeolocationServiceHelper(
+        fusedLocationClient,
+        activityLauncher
+    )
 ) {
     private lateinit var resolveLocationSettingsResultFlow: MutableSharedFlow<Result<Unit>>
     private val locationCallbacks: MutableMap<String, LocationCallback> = mutableMapOf()
@@ -57,9 +51,11 @@ class OSGeolocationController(
         try {
             // check timeout
             if (options.timeout <= 0) {
-                return Result.failure(OSLocationException.OSLocationInvalidTimeoutException(
-                    message = "Timeout needs to be a positive value."
-                ))
+                return Result.failure(
+                    OSLocationException.OSLocationInvalidTimeoutException(
+                        message = "Timeout needs to be a positive value."
+                    )
+                )
             }
 
             val checkResult: Result<Unit> =
@@ -69,7 +65,7 @@ class OSGeolocationController(
                     checkResult.exceptionOrNull() ?: NullPointerException()
                 )
             } else {
-                val location = getCurrentLocation(options)
+                val location = helper.getCurrentLocation(options)
                 return Result.success(location.toOSLocationResult())
             }
         } catch (exception: Exception) {
@@ -113,7 +109,11 @@ class OSGeolocationController(
      * @param watchId a unique id identifying the watch
      * @return Flow in which location updates will be emitted
      */
-    fun addWatch(activity: Activity, options: OSLocationOptions, watchId: String): Flow<Result<List<OSLocationResult>>> = callbackFlow {
+    fun addWatch(
+        activity: Activity,
+        options: OSLocationOptions,
+        watchId: String
+    ): Flow<Result<List<OSLocationResult>>> = callbackFlow {
 
         try {
             val checkResult: Result<Unit> =
@@ -125,12 +125,7 @@ class OSGeolocationController(
             } else {
                 locationCallbacks[watchId] = object : LocationCallback() {
                     override fun onLocationResult(location: LocationResult) {
-                        if (watchIdsBlacklist.contains(watchId)) {
-                            // received a location update but watch id is in blacklist, so the location updates should be removed
-                            val cleared = clearWatch(watchId, addToBlackList = false)
-                            if (cleared) {
-                                watchIdsBlacklist.remove(watchId)
-                            }
+                        if (checkWatchInBlackList(watchId)) {
                             return
                         }
                         val locations = location.locations.map { currentLocation ->
@@ -138,8 +133,9 @@ class OSGeolocationController(
                         }
                         trySend(Result.success(locations))
                     }
+                }.also {
+                    helper.requestLocationUpdates(options, it)
                 }
-                requestLocationUpdates(options, watchId)
             }
         } catch (exception: Exception) {
             Log.d(LOG_TAG, "Error requesting location updates: ${exception.message}")
@@ -159,46 +155,6 @@ class OSGeolocationController(
     fun clearWatch(id: String): Boolean = clearWatch(id, addToBlackList = true)
 
     /**
-     * Obtains a fresh device location.
-     * @param options location request options to use
-     * @return Location object representing the location
-     */
-    private suspend fun getCurrentLocation(options: OSLocationOptions): Location {
-
-        val locationRequest = CurrentLocationRequest.Builder()
-            .setPriority(if (options.enableHighAccuracy) Priority.PRIORITY_HIGH_ACCURACY else Priority.PRIORITY_BALANCED_POWER_ACCURACY)
-            .setMaxUpdateAgeMillis(options.maximumAge)
-            .setDurationMillis(options.timeout)
-            .build()
-
-        return fusedLocationClient.getCurrentLocation(
-            locationRequest,
-            null
-        ).await()
-    }
-
-    /**
-     * Requests updates of device location.
-     *
-     * Locations returned in callback associated with watchId
-     * @param options location request options to use
-     * @param watchId the id for this location request
-     */
-    private fun requestLocationUpdates(options: OSLocationOptions, watchId: String) {
-        val locationRequest = LocationRequest.Builder(options.timeout).apply {
-            setMaxUpdateAgeMillis(options.maximumAge)
-            setPriority(if (options.enableHighAccuracy) Priority.PRIORITY_HIGH_ACCURACY else Priority.PRIORITY_BALANCED_POWER_ACCURACY)
-            if (options.minUpdateInterval != null) {
-                setMinUpdateIntervalMillis(options.minUpdateInterval)
-            }
-        }.build()
-
-        locationCallbacks[watchId]?.let {
-            fusedLocationClient.requestLocationUpdates(locationRequest, it, Looper.getMainLooper())
-        }
-    }
-
-    /**
      * checks if all preconditions for retrieving location are met
      * @param activity the Android activity from which the location request is being triggered
      * @param options location request options to use
@@ -209,13 +165,13 @@ class OSGeolocationController(
         options: OSLocationOptions,
         isSingleLocationRequest: Boolean
     ): Result<Unit> {
-        val playServicesResult = checkGooglePlayServicesAvailable(activity)
+        val playServicesResult = helper.checkGooglePlayServicesAvailable(activity)
         if (playServicesResult.isFailure) {
             return Result.failure(playServicesResult.exceptionOrNull() ?: NullPointerException())
         }
 
         resolveLocationSettingsResultFlow = MutableSharedFlow()
-        val locationSettingsChecked = checkLocationSettings(
+        val locationSettingsChecked = helper.checkLocationSettings(
             activity,
             options,
             interval = if (isSingleLocationRequest) 0 else options.timeout
@@ -229,81 +185,6 @@ class OSGeolocationController(
     }
 
     /**
-     * Checks if location is on, as well as other conditions for retrieving device location
-     * @param activity the Android activity from which the location request is being triggered
-     * @param options location request options to use
-     * @param interval interval for requesting location updates; use 0 if meant to retrieve a single location
-     * @return true if location was checked and is on, false if it requires user to resolve issue (e.g. turn on location)
-     *          If false, the result is returned in `resolveLocationSettingsResultFlow`
-     * @throws [OSLocationException.OSLocationSettingsException] if an error occurs that is not resolvable by user
-     */
-    private suspend fun checkLocationSettings(
-        activity: Activity,
-        options: OSLocationOptions,
-        interval: Long
-    ): Boolean {
-
-        val request = LocationRequest.Builder(
-            if (options.enableHighAccuracy) Priority.PRIORITY_HIGH_ACCURACY else Priority.PRIORITY_BALANCED_POWER_ACCURACY,
-            interval
-        ).build()
-
-        val builder = LocationSettingsRequest.Builder()
-        builder.addLocationRequest(request)
-        val client = LocationServices.getSettingsClient(activity)
-
-        try {
-            client.checkLocationSettings(builder.build()).await()
-            return true
-        } catch (e: ResolvableApiException) {
-
-            // Show the dialog to enable location by calling startResolutionForResult(),
-            // and then handle the result in onActivityResult
-            val resolutionBuilder: IntentSenderRequest.Builder = IntentSenderRequest.Builder(e.resolution)
-            val resolution: IntentSenderRequest = resolutionBuilder.build()
-
-            activityLauncher.launch(resolution)
-        } catch (e: Exception) {
-            throw OSLocationException.OSLocationSettingsException(
-                message = "There is an error with the location settings.",
-                cause = e
-            )
-        }
-        return false
-    }
-
-    /**
-     * Checks if the device has google play services, required to use [FusedLocationProviderClient]
-     * @param activity the Android activity from which the location request is being triggered
-     * @return true if google play services is available, false otherwise
-     */
-    private fun checkGooglePlayServicesAvailable(activity: Activity): Result<Unit> {
-        val googleApiAvailability = GoogleApiAvailability.getInstance()
-        val status = googleApiAvailability.isGooglePlayServicesAvailable(activity)
-
-        return if (status != ConnectionResult.SUCCESS) {
-            if (googleApiAvailability.isUserResolvableError(status)) {
-                googleApiAvailability.getErrorDialog(activity, status, 1)?.show()
-                Result.failure(
-                    OSLocationException.OSLocationGoogleServicesException(
-                        resolvable = true,
-                        message = "Google Play Services error user resolvable."
-                    )
-                )
-            } else {
-                Result.failure(
-                    OSLocationException.OSLocationGoogleServicesException(
-                        resolvable = false,
-                        message = "Google Play Services error."
-                    )
-                )
-            }
-        } else {
-            Result.success(Unit)
-        }
-    }
-
-    /**
      * clears a watch by removing its location update request
      * @param id the watch id
      * @param addToBlackList whether or not the watch id should go in blacklist if not found
@@ -312,7 +193,7 @@ class OSGeolocationController(
     private fun clearWatch(id: String, addToBlackList: Boolean): Boolean {
         val locationCallback = locationCallbacks.remove(key = id)
         return if (locationCallback != null) {
-            fusedLocationClient.removeLocationUpdates(locationCallback)
+            helper.removeLocationUpdates(locationCallback)
             true
         } else {
             if (addToBlackList) {
@@ -324,11 +205,29 @@ class OSGeolocationController(
         }
     }
 
+    /**
+     * Checks if the current watch is in the blacklist
+     *
+     * If the watch is in the blacklist, location updates for that watch should be removed.
+     * @param watchId the unique id of the watch
+     * @return true if watch is in blacklist, false otherwise
+     */
+    private fun checkWatchInBlackList(watchId: String): Boolean {
+        if (watchIdsBlacklist.contains(watchId)) {
+            val cleared = clearWatch(watchId, addToBlackList = false)
+            if (cleared) {
+                watchIdsBlacklist.remove(watchId)
+            }
+            return true
+        }
+        return false
+    }
+
     private fun Location.toOSLocationResult(): OSLocationResult = OSLocationResult(
         latitude = this.latitude,
         longitude = this.longitude,
         altitude = this.altitude,
-        accuracy =  this.accuracy,
+        accuracy = this.accuracy,
         altitudeAccuracy = if (OSGeolocationBuildConfig.getAndroidSdkVersionCode() >= Build.VERSION_CODES.O) this.verticalAccuracyMeters else null,
         heading = this.bearing,
         speed = this.speed,
